Correct the g^4 terms (DRbar, 2-loop, xi=1) in the MSSM and NMSSM.
The old g^4 terms were taken from SARAH.  But SARAH used the anomalous
dimensions of the Higgs superfields, which must not be used in the
beta functions of the VEVs.  Instead, the anomalous dimensions of the
scalar Higgs component fields must be used.  Marcus Sperling has
calculated them in the MSSM and NMSSM (they are the same in both
models due to the same gauge group).