remove todo item and empty if statement body
Note: We currently do not set xiS in the NMSSM high-scale bcs.
This is ok, because

 1) in the Z3 conserving case xiS will always be zero and

 2) in the Z3 violating case xiS is an EWSB output parameter.

But if we (at some point) decide to provide a different set of EWSB
output parameters which does not include xiS in the Z3 violating case,
we have to set it properly at the GUT scale.