addressing Reviewer #1, comment 1)
Clarify the definiton of M_{SUSY}.  This is the scale where we input
NMSSM parameters at, if a -1 entry in the QEXTPAR block is given.

Note, Reviewer #1, comment 1) contains a misunderstanding of the
meaning MODSEL entry 12.  According to SLHA (hep-ph/0311123), MODSEL
entry 12 defines the (max.) output scale for the running parameters.
It does not set the value of M_{SUSY}.