renaming --BmuOverCosBetaSinBeta= cmd line option for consistency
In the NMSSM manual we're using m_3^2 everywhere, not Bmu.  The
available command line options should reflect the conventions from the
manual.

Warning: This is an command line interface change.  However, we still
allow the use of the --BmuOverCosBetaSinBeta= option for backward
compatibility (but print a warning that the option is deprecated).