Add initial calculation of (pole) Z mass
Note that the method used in the MSSM, where the
self-energy is calculated by subtracting the running
mass from the experimental value, does not work as
currently implemented.  This is because if the VEVs
satisfy the EWSB conditions, the running mass trivially
cancels out in the expression and we end
up with MZ = 91.1876 as the predicted mass always.