Do not set model error status in fine tuning calculation
Problems like tachyons can appear in the VEV iteration
during the tuning calculation, but these at worst just
mean that the computed fine tuning is inaccurate or
wrong, and should not flag the whole point as being
invalid.  Instead errors in the tuning calculation
should be checked and if found, we should return an
invalid value.