Use forward/backward derivatives as well in Jacobian calculation
in the event that the estimated error obtained using central
differences is large.  The idea here is that there is a chance
that this could be occurring because we are at a boundary in
parameter space, so maybe better results can be obtained
considering only variations away from the boundary.